added security to person endpoints

+++ src/main/java/com/dat250/FeedApp/controller/PersonController.java
@@ -5,8 +5,10 @@ import com.dat250.FeedApp.service.PersonService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -31,19 +33,33 @@ public class PersonController {
         return personService.getPerson(personId);
     }
 
-    @PostMapping("/people")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Person createPerson(@Validated @RequestBody Person person) {
-        return personService.createPerson(person);
-    }
+    /*    Replaced by PublicUsersController
+        @PostMapping("/people")
+        @ResponseStatus(HttpStatus.CREATED)
+        public Person createPerson(@Validated @RequestBody Person person) {
+            return personService.createPerson(person);
+        }
+     */
 
     @PutMapping("/people/{personId}")
-    public Person updatePerson(@PathVariable Long personId, @Validated @RequestBody Person personRequest) {
-        return personService.updatePerson(personId, personRequest);
+    public Person updatePerson(@PathVariable Long personId, @Validated @RequestBody Person personRequest
+            , @AuthenticationPrincipal final Person user) {
+        if (personId.equals(user.getPersonId()) || user.isAdmin()) {
+            return personService.updatePerson(personId, personRequest);
+        } else {
+            System.err.println("access denied for user: " + user.getName());
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Person: " + user.getName() + " is an imposter");
+        }
+
     }
 
     @DeleteMapping("/people/{personId}")
-    public ResponseEntity<?> deletePerson(@PathVariable Long personId) {
-        return personService.deletePerson(personId);
+    public ResponseEntity<?> deletePerson(@PathVariable Long personId, @AuthenticationPrincipal final Person user) {
+        if (personId.equals(user.getPersonId()) || user.isAdmin()) {
+            return personService.deletePerson(personId);
+        } else {
+            System.err.println("access denied for user: " + user.getName());
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Person: " + user.getName() + " is an imposter");
+        }
     }
 }
